refactor: add  style.js
