+++ src/pages/Trainee/Components/Table/tableStyle.js
@@ -1,0 +1,20 @@
+const useStyles = (theme) => ({
+  table: {
+    minWidth: 600,
+    spacing: 8,
+  },
+  tableCell: {
+    color: 'darkgray',
+  },
+  header: {
+    color: 'grey',
+  },
+  row: {
+    '&:nth-of-type(odd)': {
+      backgroundColor: theme.palette.action.hover,
+    },
+    cursor: 'pointer',
+  },
+});
+
+export { useStyles };
